Fix complet du système de création d'articles - Résolution des erreurs TypeError et implémentation du système d'upload d'images
 Corrections apportées :
- Fix TypeError dans handleImageUpload() : conversion (int) pour Database::lastInsertId()
- Fix TypeError dans Article::create() : conversion (int) pour Database::lastInsertId()
- Fix TypeError dans Auth::logActivity() : paramètres dans le bon ordre
- Implémentation système d'upload d'images direct avec validation
- Ajout champ status (brouillon/publié/archivé) avec published_at dynamique
- Système intelligent cover_image_id (priorité jeu > upload direct)
- Validation JavaScript côté client pour tous les champs requis
- Correction drag & drop sur tous les modules de l'éditeur
- Système de flash messages fonctionnel
- API endpoints pour récupération jeux et médias

 Fonctionnalités opérationnelles :
- Création d'articles complète avec éditeur modulaire
- Upload sécurisé d'images de couverture
- Validation robuste côté client et serveur
- Log d'activité fonctionnel
- Interface utilisateur intuitive et responsive

+++ app/models/Article.php
@@ -107,7 +107,7 @@ class Article
                 $data['featured_position'] ?? null
             ];
             
-            return Database::execute($sql, $params) ? Database::lastInsertId() : null;
+            return Database::execute($sql, $params) ? (int)Database::lastInsertId() : null;
         } catch (Exception $e) {
             error_log("Erreur création article: " . $e->getMessage());
             return null;
